issue_12 fix various
