+++ parsinggettingthingsdone/main.py
@@ -9,7 +9,7 @@ import sys
 sys.path.insert(0, '../parsinggettingthingsdone')
 
 from datetime import datetime
-from physical import CSV
+from formatter import CSV
 from physical import Directory
 from physical import Filesystem
